moves progress computation functions to seperate module
